+++ src/OutlookGoogleCalendarSync/SettingsStore/Proxy.cs
@@ -24,7 +24,7 @@ namespace OutlookGoogleCalendarSync.SettingsStore {
         }
 
         public String DefaultBrowserAgent {
-            get { return "Mozilla / 5.0(Windows NT 6.1; WOW64; Trident / 7.0; rv: 11.0) like Gecko"; }
+            get { return "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/103.0.0.0 Safari/537.36"; }
         }
         private void setDefaults() {
             //Default values for new class