+++ src/OutlookGoogleCalendarSync/Sync/Calendar.cs
@@ -390,11 +390,14 @@ namespace OutlookGoogleCalendarSync.Sync {
                             DateTime checkDates = ai.Start;
                             checkDates = ai.End;
                         } catch (System.Runtime.InteropServices.COMException ex) {
-                            if (OGCSexception.GetErrorCode(ex, 0x0000FFFF) == "0x00004005") { //You must specify a time/hour
+                            if (OGCSexception.GetErrorCode(ex) == "0x80040305" || //Your server administrator has limited the number of items you can open simultaneously.
+                                OGCSexception.GetErrorCode(ex, 0x000FFFFF) == "0x00040115") //Network problems are preventing connection to Microsoft Exchange.
+                            {
+                                Forms.Main.Instance.Console.UpdateWithError("Cannot continue synchronising.", ex);
+                                return SyncResult.AutoRetry;
+                            } else if (OGCSexception.GetErrorCode(ex, 0x0000FFFF) == "0x00004005") { //You must specify a time/hour
                                 skipCorruptedItem(ref outlookEntries, outlookEntries[o], ex.Message);
                             } else {
-                                //"Your server administrator has limited the number of items you can open simultaneously."
-                                //Once we have the error code for above message, need to abort sync - and suggest using cached Exchange mode
                                 OGCSexception.Analyse("Calendar item does not have a proper date range - cannot sync it. ExchangeMode=" +
                                     OutlookOgcs.Calendar.Instance.IOutlook.ExchangeConnectionMode().ToString(), ex);
                                 skipCorruptedItem(ref outlookEntries, outlookEntries[o], ex.Message);
@@ -467,15 +470,18 @@ namespace OutlookGoogleCalendarSync.Sync {
                     OutlookOgcs.Calendar.Instance.ReclaimOrphanCalendarEntries(ref outlookEntries, ref googleEntries);
                     if (Sync.Engine.Instance.CancellationPending) return SyncResult.UserCancelled;
 
-                    if (this.Profile.AddColours || this.Profile.SetEntriesColour) OutlookOgcs.Calendar.Categories.ValidateCategories();
-                    if (this.Profile.ColourMaps.Count > 0) {
-                        this.Profile.ColourMaps.ToList().ForEach(c => {
-                            if (OutlookOgcs.Calendar.Categories.OutlookColour(c.Key) == null) {
-                                if (OgcsMessageBox.Show("There is a problem with your colour mapping configuration.\r\nColours may not get synced as intended.\r\nReview maps now for missing Outlook colours?",
-                                    "Invalid colour map", System.Windows.Forms.MessageBoxButtons.YesNo, System.Windows.Forms.MessageBoxIcon.Error) == DialogResult.Yes)
-                                    new Forms.ColourMap().ShowDialog();
-                            }
-                        });
+                    if (this.Profile.AddColours || this.Profile.SetEntriesColour) {
+                        OutlookOgcs.Calendar.Categories.ValidateCategories();
+                     
+                        if (this.Profile.ColourMaps.Count > 0) {
+                            this.Profile.ColourMaps.ToList().ForEach(c => {
+                                if (OutlookOgcs.Calendar.Categories.OutlookColour(c.Key) == null) {
+                                    if (OgcsMessageBox.Show("There is a problem with your colour mapping configuration.\r\nColours may not get synced as intended.\r\nReview maps now for missing Outlook colours?",
+                                        "Invalid colour map", System.Windows.Forms.MessageBoxButtons.YesNo, System.Windows.Forms.MessageBoxIcon.Error) == DialogResult.Yes)
+                                        new Forms.ColourMap().ShowDialog();
+                                }
+                            });
+                        }
                     }
 
                     //Sync