+++ docs/Release Notes.md
@@ -1,6 +1,38 @@
 | Release Notes |
 |:---|
 
+# v2.9.6.0 - Alpha
+
+**Enhancements**
+- Warn if multiple instances of OGCS are running with the same configuration
+- Cater for Kyiv time zone rename (from Kiev)
+- Browser agent version updated
+- Display error detail if OGCS unable to communicate with Google
+- Abort sync if Exchange connection becomes unavailable
+- Better UI for colour map configuration
+
+**Bugfix**
+- Error when switching Profiles
+- Use correct Find/Replace regular expression from Profile
+- Resolve "Precondition Failed", error [412] on sync to Google
+- Fix null reference errors on sync to Google
+- Don't validate colour mappings if no longer syncing colours
+- Improved handling of startup registry keys
+
+# v2.9.5.0 - Alpha
+
+**Enhancements**
+- New option to set synced items as 'public'
+- Categories now properly populating for Alternate Mailbox
+- New option to set OGCS to startup for all Windows users, not just current
+- Show warning that filtered out items may sync as "duplicates"
+
+**Bugfix**
+- Don't touch Outlook attendee response status, if they are Google organiser
+- Fixed error accessing Outlook item that has been modified
+- Fixed error when syncing colours, but none set on Outlook appointment
+- Handle being disconnected from Outlook when retrieving categories
+
 # v2.9.4.0 - Alpha
 
 **Enhancements**